refactor: move containers to components folder
